Refactor cone-of-influence analysis into separate class
